+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## 1.1.4 - 2018-05-07
 
-* `DIRECTORY_SEPARATOR` was added to the paths to avoid incompatibilities with Windows paths.
+* Added `Eliasis\Framework\View::setHash()` method.
+
+* Added method `View::setHash()` to fixed imprecise paths in Windows.
 
 ## 1.1.3 - 2018-04-17